+++ markdown-toc-pkg.el
@@ -1,4 +1,4 @@
-(define-package "markdown-toc" "0.1.0" "A simple TOC generator for markdown file"
+(define-package "markdown-toc" "0.1.2" "A simple TOC generator for markdown file"
   '((s "1.9.0")
     (dash "2.11.0")
     (markdown-mode "2.1")))