+++ markdown-toc.el
@@ -5,7 +5,7 @@
 ;; Maintainer: Antoine R. Dumont
 ;; URL: http://github.com/ardumont/markdown-toc
 ;; Created: 24th May 2014
-;; Version: 0.1.0
+;; Version: 0.1.2
 ;; Keywords: markdown, toc, tools,
 ;; Package-Requires: ((markdown-mode "2.1") (dash "2.11.0") (s "1.9.0"))
 
@@ -62,7 +62,7 @@
 (require 'dash)
 (require 'markdown-mode)
 
-(defconst markdown-toc--toc-version "0.1.0" "Current version installed.")
+(defconst markdown-toc--toc-version "0.1.2" "Current version installed.")
 
 (defgroup markdown-toc nil
   "A simple TOC generator for markdown file."
@@ -101,12 +101,18 @@
        (-repeat it sym)
        (s-join "" it)))
 
+(defconst markdown-toc--protection-symbol "09876543214b825dc642cb6eb9a060e54bf8d69288fbee49041234567890"
+  "Implementation detail to protect some punctuation characters
+  when converting to link.")
+
 (defun markdown-toc--to-link (title)
   "Given a TITLE, return the markdown link associated."
   (format "[%s](#%s)" title
           (->> title
                downcase
+               (replace-regexp-in-string "-" markdown-toc--protection-symbol)
                (replace-regexp-in-string "[[:punct:]]" "")
+               (replace-regexp-in-string markdown-toc--protection-symbol "-")
                (s-replace " " "-"))))
 
 (defun markdown-toc--to-markdown-toc (level-title-toc-list)
@@ -118,19 +124,27 @@
                         (markdown-toc--to-link title))))
        (s-join "\n")))
 
-(defconst markdown-toc--header-toc-start
-  "<!-- markdown-toc start - Don't edit this section. Run M-x markdown-toc-generate-toc again -->")
-(defconst markdown-toc--header-toc-title
-  "**Table of Contents**")
-(defconst markdown-toc--header-toc-end
-  "<!-- markdown-toc end -->")
+(defcustom markdown-toc-header-toc-start
+  "<!-- markdown-toc start - Don't edit this section. Run M-x markdown-toc-generate-toc again -->"
+  "Beginning delimiter comment."
+  :group 'markdown-toc)
+
+(defcustom markdown-toc-header-toc-title
+  "**Table of Contents**"
+  "Title comment on TOC header."
+  :group 'mardown-toc)
+
+(defcustom markdown-toc-header-toc-end
+  "<!-- markdown-toc end -->"
+  "Ending delimiter comment."
+  :group 'mardown-toc)
 
 (defun markdown-toc--toc-already-present-p ()
   "Determine if a TOC has already been generated.
 Return the end position if it exists, nil otherwise."
   (save-excursion
     (goto-char (point-min))
-    (re-search-forward markdown-toc--header-toc-start nil t)))
+    (re-search-forward markdown-toc-header-toc-start nil t)))
 
 (defun markdown-toc--toc-start ()
   "Compute the toc's starting point."
@@ -142,7 +156,7 @@ Return the end position if it exists, nil otherwise."
   "Compute the toc's end point."
   (save-excursion
     (goto-char (point-min))
-    (re-search-forward markdown-toc--header-toc-end nil t)))
+    (re-search-forward markdown-toc-header-toc-end nil t)))
 
 (defun markdown-toc--generate-toc (toc-structure)
   "Given a TOC-STRUCTURE, compute a new toc."
@@ -153,10 +167,10 @@ Return the end position if it exists, nil otherwise."
 (defun markdown-toc--compute-full-toc (toc)
   "Given the TOC's content, compute the full toc with comments and title."
   (format "%s\n%s\n\n%s\n\n%s\n"
-          markdown-toc--header-toc-start
-          markdown-toc--header-toc-title
+          markdown-toc-header-toc-start
+          markdown-toc-header-toc-title
           toc
-          markdown-toc--header-toc-end))
+          markdown-toc-header-toc-end))
 
 (defcustom markdown-toc-user-toc-structure-manipulation-fn
   (lambda (toc-structure) toc-structure)